requirements.txt
alembic==1.8.1
asttokens==2.4.1
autopep8==1.7.0
bcrypt==4.0.0
black==22.12.0
blinker==1.5
click==8.1.3
colorama==0.4.5
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
exceptiongroup==1.2.2
executing==2.0.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
ipython==8.26.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mypy-extensions==0.4.3
parso==0.8.4
pathspec==0.10.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==2.6.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.9.1
Pygments==2.18.0
python-decouple==3.6
python-dotenv==1.0.1
RPi.GPIO==0.7.1
six==1.16.0
SQLAlchemy==1.4.40
stack-data==0.6.3
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
